Add files via upload
In this video, you will learn the basics of #HTML and #CSS from the beginning. This HTML and CSS course is best for Beginners to start learning web design and web development. In this HTML and CSS Beginners tutorial, we will discuss all the important HTML tags, and CSS properties to design the website with some examples and explanations. We will also learn about animation using CSS.
Please watch this video till the end.